split up the Dune Client class into multiple API components
